requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
openpyxl
requests
python-multipart
pydantic